+++ spec/grape/api_spec.rb
@@ -3982,7 +3982,7 @@ describe Grape::API do
       subject.format :json
       subject.get '/endpoint/options' do
         {
-          path: options[:path],
+          path: route.path,
           source_location: source.source_location
         }
       end
@@ -3991,7 +3991,7 @@ describe Grape::API do
     it 'path' do
       get '/endpoint/options'
       options = Grape::Json.parse(last_response.body)
-      expect(options['path']).to eq(['/endpoint/options'])
+      expect(options['path']).to eq('/endpoint/options(.json)')
       expect(options['source_location'][0]).to include 'api_spec.rb'
       expect(options['source_location'][1].to_i).to be > 0
     end